Disabled localStorage methods in the browser

+++ client/app/scripts/services/localStorage.js
@@ -15,96 +15,109 @@ app.service('LocalStorage', [
   var dataDir;
   Platform.ready
   .then( function() {
-    dataDir = cordova.file.dataDirectory;
+    if(!Platform.isBrowser()) {
+      dataDir = cordova.file.dataDirectory;
+    }
   });
 
   var writeFile = function(dir ,filename, data) {
+    if(Platform.isBrowser()) {
+      console.log('Cannot write files in the browser');
+    } else {
+      var write = function() {
+        var path = dataDir;
+        if(dir.length > 0) {
+          path += dir + '/';
+        }
+        $cordovaFile.writeFile(path, filename, data, true)
+        .then( function() {
+          //console.log('Successful write of ' + filename);
+        }, function(err) {
+          console.log('Error: Failed to write ' + dataDir + dir + '/' + filename + ': ' + JSON.stringify(err));
+        });
+      };
 
-    var write = function() {
+      $cordovaFile.checkDir(dataDir, dir)
+      .then( write(), function(err) {
+        if(err.code === 1) {
+          $cordovaFile.createDir(dataDir,dir,true)
+          .then( write(), function(err) {
+            console.log('Error: Failed to create directory ' + dataDir + dir + ': ' + JSON.stringify(err));
+          });
+        }
+        else {
+          console.log('Error: There was an error checking the directory ' + dataDir + dir + ': ' + JSON.stringify(err));
+        }
+      });
+    }
+  };
+
+  var readFiles = function(dir, cb) {
+    if(Platform.isBrowser()) {
+      console.log('Cannot read files in the browser');
+    } else {
       var path = dataDir;
       if(dir.length > 0) {
         path += dir + '/';
       }
-      $cordovaFile.writeFile(path, filename, data, true)
-      .then( function() {
-        //console.log('Successful write of ' + filename);
-      }, function(err) {
-        console.log('Error: Failed to write ' + dataDir + dir + '/' + filename + ': ' + JSON.stringify(err));
-      });
-    };
 
-    $cordovaFile.checkDir(dataDir, dir)
-    .then( write(), function(err) {
-      if(err.code === 1) {
-        $cordovaFile.createDir(dataDir,dir,true)
-        .then( write(), function(err) {
-          console.log('Error: Failed to create directory ' + dataDir + dir + ': ' + JSON.stringify(err));
-        });
-      }
-      else {
-        console.log('Error: There was an error checking the directory ' + dataDir + dir + ': ' + JSON.stringify(err));
-      }
-    });
-  };
-
-  var readFiles = function(dir, cb) {
-    var path = dataDir;
-    if(dir.length > 0) {
-      path += dir + '/';
-    }
-
-    var done = 0;
-    var total = 0;
-    var files = [];
-    var handle = function(err,res) {
-      done++;
-      if(res) {
-        files.push(res);
-      }
-      if(done === total) {
-        console.log('Success: Read the directory!'); 
-        cb(null, files);
-      }
-    };
-
-    window.resolveLocalFileSystemURL(path, function(fileSystem) {
-      var dirReader = fileSystem.createReader();
-      dirReader.readEntries( function(res) {
-        total = res.length;
-        for( var i = 0; i < res.length; i++) {
-          var name = res[i].fullPath.split('/');
-          readFile(dir, String(name[name.length -1]), handle);
+      var done = 0;
+      var total = 0;
+      var files = [];
+      var handle = function(err,res) {
+        done++;
+        if(res) {
+          files.push(res);
         }
-      }, function(err) {
-        console.log('Error: Failed to read the directory! ' + JSON.stringify(err));
-        cb(err);
+        if(done === total) {
+          console.log('Success: Read the directory!'); 
+          cb(null, files);
+        }
+      };
+
+      window.resolveLocalFileSystemURL(path, function(fileSystem) {
+        var dirReader = fileSystem.createReader();
+        dirReader.readEntries( function(res) {
+          total = res.length;
+          for( var i = 0; i < res.length; i++) {
+            var name = res[i].fullPath.split('/');
+            readFile(dir, String(name[name.length -1]), handle);
+          }
+        }, function(err) {
+          console.log('Error: Failed to read the directory! ' + JSON.stringify(err));
+          cb(err);
+        });
       });
-    });
+    }
   };
 
   var readFile = function(dir, filename, cb) {
-    var path = dataDir;
-    if(dir.length > 0) {
-      path += dir + '/';
-    }
-    $cordovaFile.readAsText(path, filename)
-    .then( function(text) {
-      var types = filename.split('.');
-      var type = types[types.length - 1];
-
-      var res;
-      if( type === 'json' ) {
-        res = JSON.parse(text);
-      } else {
-        res = text;
+    if(Platform.isBrowser()) {
+      console.log('Cannot read files in the browser');
+    } else {
+      var path = dataDir;
+      if(dir.length > 0) {
+        path += dir + '/';
       }
+      $cordovaFile.readAsText(path, filename)
+      .then( function(text) {
+        var types = filename.split('.');
+        var type = types[types.length - 1];
 
-      cb(null,res);
+        var res;
+        if( type === 'json' ) {
+          res = JSON.parse(text);
+        } else {
+          res = text;
+        }
 
-    }, function(err) {
-      console.log('Error: Failed to read file ' + filename + ': ' + JSON.stringify(err));
-      cb(err.message);
-    });
+        cb(null,res);
+
+      }, function(err) {
+        console.log('Error: Failed to read file ' + filename + ': ' + JSON.stringify(err));
+        cb(err.message);
+      });
+    }
   };
 
    var getFileName = function(partial) {
@@ -113,7 +126,9 @@ app.service('LocalStorage', [
 
    //TODO Use native encryption on Android and keychain on iOS
    var secureWrite = function(key, value) {
-
+    if(Platform.isBrowser()) {
+      console.log('Cannot write files in the browser');
+    } else {
       var val = JSON.stringify(value);
       var encrypted = CryptoJS.AES.encrypt(val, key).toString();
 
@@ -123,10 +138,13 @@ app.service('LocalStorage', [
 
       writeFile('', file, encrypted);
       //console.log('Successfully wrote ' + JSON.stringify(value, 'utf8', '  ') + ' to ' + key + ' (' + file + ')');
+     }
    };
 
    var secureRead = function(key, cb) {
-
+    if(Platform.isBrowser()) {
+      console.log('Cannot read files in the browser');
+    } else {
       var file = getFileName(CryptoJS.SHA256(key));
  //     console.log('Trying to read ' + file);
 
@@ -146,10 +164,14 @@ app.service('LocalStorage', [
          cb(err.message, null);
         }
       });
+    }
    };
 
    //TODO Ensure the data gets overwritten at the deletion address
    var secureDelete = function(key) {
+    if(Platform.isBrowser()) {
+      console.log('Cannot delete files in the browser');
+    } else {
       var file = getFileName(CryptoJS.SHA256(key));
       //console.log('Trying to delete ' + file);
 
@@ -159,15 +181,20 @@ app.service('LocalStorage', [
       }, function (err) {
          console.log('Error deleting file: '+ err.message);
       });
+    }
    };
 
    var deleteFile = function(dir, file) {
+    if(Platform.isBrowser()) {
+      console.log('Cannot delete files in the browser');
+    } else {
       $cordovaFile.removeFile(dataDir + dir, file)
       .then( function (success) {
          console.log('Successfully deleted file!'+ JSON.stringify(success.fileRemoved.name));
       }, function (err) {
          console.log('Error deleting file: '+ JSON.stringify(err));
       });
+    }
    };
 
   return {
